ネイティブライブラリのパッケージの構成の変更

+++ PublishPackage.ps1
@@ -8,19 +8,21 @@ if (-not (Test-Path $localNugetPackageDirectory -PathType Container)) {
 }
 
 foreach ($projectDirectory in Get-ChildItem $baseDirectory -Directory | Where-Object {$_.Name -match "^Palmtree\."}) {
-    $nugetPackageFiles = Get-ChildItem (Join-Path (Join-Path $projectDirectory.FullName "bin") "Release") -File | Where-Object {$_.Name -match "\.nupkg$"}
-    foreach ($nugetPackageFile in $nugetPackageFiles) {
-        $command = "nuget"
-        $args = "add `"" + $nugetPackageFile.FullName + "`" -Source `"" + $localNugetPackageDirectory + "`""
-
-        $process = Start-Process -FilePath $command -ArgumentList $args -NoNewWindow -PassThru
-        $handle = $process.Handle # おまじない https://stackoverflow.com/questions/10262231/obtaining-exitcode-using-start-process-and-waitforexit-instead-of-wait
-        $process.WaitForExit()
-        if ($process.ExitCode -ne 0) {
-            Write-Warning ("nuget の実行に失敗しました。: 終了コード: " + $process.ExitCode)
-        }
-        else {
-            Copy-Item $nugetPackageFile $localNugetPackageDirectory
-        }
+    $projectReleasePath = Join-Path $projectDirectory.FullName "bin/Release"
+    if (Test-Path $projectReleasePath -PathType Container) {
+        $nugetPackageFiles = Get-ChildItem $projectReleasePath -File | Where-Object {$_.Name -match "\.nupkg$"}
+        foreach ($nugetPackageFile in $nugetPackageFiles) {
+            $command = "nuget"
+            $args = "add `"" + $nugetPackageFile.FullName + "`" -Source `"" + $localNugetPackageDirectory + "`""
+            $process = Start-Process -FilePath $command -ArgumentList $args -NoNewWindow -PassThru
+            $handle = $process.Handle # おまじない https://stackoverflow.com/questions/10262231/obtaining-exitcode-using-start-process-and-waitforexit-instead-of-wait
+            $process.WaitForExit()
+            if ($process.ExitCode -ne 0) {
+                Write-Warning ("nuget の実行に失敗しました。: 終了コード: " + $process.ExitCode)
+            }
+            else {
+                Copy-Item $nugetPackageFile $localNugetPackageDirectory
+            }
+       }
     }
 }
